v4l2: fix transform_caps callback at gstv4l2transform
The GstConvert transform were tested in the past with the
assumption that both source and sync pads would support
the same formats.

Probably due to that, it was never noticed that the
transform_caps actually from the source direction are
actually validating the sink direction, and vice versa,
as described at GstBaseTransformClass description.

With a recent Kernel patch added bayer formats only for
the V4L2 camera type, this problem became evident.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>